requirements.txt
amqp==2.1.4
beautifulsoup4==4.5.3
billiard==3.5.0.2
bs4==0.0.1
celery==4.0.2
certifi==2016.9.26
Django==1.10.4
future==0.16.0
kombu==4.0.2
python-telegram-bot==5.3.0
pytz==2016.10
requests==2.12.4
urllib3==1.19.1
vine==1.1.3
